{ref}[Ini]: Switched all page-loading function to page-specific files. | {fix}[Run]: Modified fav loading loop to count the fav models in file.
