+++ src/gui/pages/MonkeySCPage.cpp
@@ -1,0 +1,3 @@
+//
+// Created by alber on 18/06/2024.
+//
